Re-organization of main_wf scripts, where each processing step has its own main_wf.py file within its package folder.
