+++ rabies/utils.py
@@ -1,0 +1,50 @@
+
+######################
+#FUNCTIONS TO READ WORKFLOW GRAPH
+######################
+import pathlib
+
+def fill_split_dict(d, output_bold, split_name, split_dict, keys, node_dict, match_targets):
+    if isinstance(d, dict):
+        for key in list(d.keys()):
+            fill_split_dict(d[key], output_bold, split_name, split_dict, keys+[key], node_dict, match_targets)
+    else:
+        f = d.result.outputs.get()[output_bold]
+        split = pathlib.Path(f).name.rsplit(".nii")[0]
+        split_name.append(split)
+        split_dict[split]={}
+        target_list = list(match_targets.keys())
+        for target in target_list:
+            [unit, output] = match_targets[target]
+            node = retrieve_node(node_dict[unit], keys)
+            split_dict[split][target] = node.result.outputs.get()[output]
+        
+def retrieve_node(d, keys):
+    if isinstance(d, dict):
+        return retrieve_node(d[keys[0]], keys[1:])
+    else:
+        return d
+
+
+def get_workflow_dict(workflow_file):
+    import pickle
+    with open(workflow_file, 'rb') as handle:
+        graph = pickle.load(handle)
+    
+    node_list = list(graph.nodes)
+    node_dict = {}
+    for node in node_list:
+        key_l = [node.fullname]+node.parameterization
+        fill_node_dict(node_dict, key_l, node)
+    return node_dict
+
+
+def fill_node_dict(d, key_l, e):
+    if len(key_l)>0:
+        key = key_l[0]
+        if not (key in list(d.keys())):
+            d[key] = {}
+        d[key] = fill_node_dict(d[key], key_l[1:], e)
+        return d
+    else:
+        return e
